✔️ Fix small problems
1- The style was applied glabaly on some buttons and inputs
2- Deletion blogs now require user confirmation
3- Deletion of blog in blog page now redirect to home page
4- Reduced unneccessary code in /helper/config.js
5- Added note TOFIX to notifications

+++ ui-admin/src/components/BlogCard.jsx
@@ -1,11 +1,18 @@
 import BlogService from "../services/blog.js";
+import { useNavigate } from "@solidjs/router";
 
 function BlogCard({ blog, mutate }) {
+  const navigate = useNavigate();
   const handleUpdate = () => {
-    window.location.href = `/blogs/update/${blog.id}`;
+    navigate(`/blogs/update/${blog.id}`);
+    // window.location.href = `/blogs/update/${blog.id}`;
   };
 
   const handleDelete = () => {
+    const confimChoice = window.confirm(
+      `Are you sure you want to delete This blog titled : ${blog.title}`
+    );
+    if (!confimChoice) return;
     BlogService.remove(blog.id);
     mutate((array) => array.filter((item) => item.id != blog.id));
   };
